refactor: handle error in review form submission and add action state
