+++ src/actions/create-review.action.ts
@@ -2,23 +2,41 @@
 
 import { revalidateTag } from "next/cache";
 
-export async function createReviewAction(formData: FormData) {
+export async function createReviewAction(_: any, formData: FormData) {
   const movieId = formData.get("movieId")?.toString();
   const author = formData.get("author")?.toString();
   const content = formData.get("content")?.toString();
   const rating = formData.get("rating")?.toString();
 
-  if (!movieId || !author || !content) return;
+  if (!movieId || !author || !content) {
+    return {
+      status: false,
+      error: `Username and Review content cannot be empty.`,
+    };
+  }
 
   try {
-    await fetch(`${process.env.NEXT_PUBLIC_API_SERVER_URL}/review`, {
-      method: "POST",
-      body: JSON.stringify({ movieId, author, content, rating }),
-    });
+    const res = await fetch(
+      `${process.env.NEXT_PUBLIC_API_SERVER_URL}/review/aa`,
+      {
+        method: "POST",
+        body: JSON.stringify({ movieId, author, content, rating }),
+      }
+    );
+
+    if (!res.ok) {
+      throw new Error(`POST request failed: ${res.statusText}`);
+    }
 
     revalidateTag(`review-${movieId}`);
+    return {
+      status: true,
+      error: "'",
+    };
   } catch (error) {
-    console.error(error);
-    return;
+    return {
+      status: false,
+      error: `Submission failed : ${error}`,
+    };
   }
 }
