+++ src/app/movie/[id]/page.tsx
@@ -2,8 +2,8 @@ import Image from "next/image";
 import { css, cva } from "../../../../styled-system/css";
 import { MovieData, ReviewData } from "@/app/type";
 import { notFound } from "next/navigation";
-import { createReviewAction } from "@/actions/create-review.action";
 import ReviewItem from "@/components/ReviewItem";
+import ReviewEditor from "@/components/ReviewEditor";
 
 interface IProps {
   params: {
@@ -80,52 +80,6 @@ async function MovieDetail({ movieId }: { movieId: string }) {
   );
 }
 
-function ReviewEditor({ movieId }: { movieId: string }) {
-  return (
-    <div>
-      <h3
-        className={css({
-          fontSize: "18px",
-          fontWeight: 600,
-          marginTop: "40px",
-        })}
-      >
-        Audience Reviews
-      </h3>
-      <form action={createReviewAction} className={formStyle}>
-        <input type="text" defaultValue={movieId} name="movieId" hidden />
-        <input
-          required
-          type="text"
-          placeholder="Username"
-          name="author"
-          className={inputStyle({ type: "name" })}
-        />
-        {/* <input
-          required
-          type="number"
-          placeholder="Rate the movie (1~5)"
-          min={1}
-          max={5}
-          name="rating"
-          className={inputStyle({ type: "name" })}
-        /> */}
-        <textarea
-          required
-          placeholder="Add a review"
-          name="content"
-          rows={3}
-          className={inputStyle({ type: "review" })}
-          maxLength={200}
-        />
-        <button type="submit" className={buttonStyle}>
-          Submit
-        </button>
-      </form>
-    </div>
-  );
-}
-
 async function ReviewList({ movieId }: { movieId: string }) {
   const res = await fetch(
     `${process.env.NEXT_PUBLIC_API_SERVER_URL}/review/movie/${movieId}`,
@@ -185,46 +139,4 @@ const posterStyle = css({
   position: "relative",
   width: "300px",
   aspectRatio: "1/1.5",
-});
-
-const formStyle = css({
-  width: "100%",
-  display: "flex",
-  flexDirection: "column",
-  gap: "4",
-  marginTop: "10px",
-});
-
-const inputStyle = cva({
-  base: {
-    border: `solid 1px var(--gray-primary)`,
-    borderRadius: "3px",
-    width: "100%",
-    padding: "5px",
-  },
-  variants: {
-    type: {
-      name: {
-        maxWidth: "200px",
-        height: "35px",
-      },
-      review: {
-        resize: "none",
-      },
-    },
-  },
-});
-
-const buttonStyle = css({
-  display: "inline-block",
-  maxWidth: "min-content",
-  alignSelf: "flex-end",
-  fontSize: "14px",
-  padding: "10px 25px",
-  backgroundColor: "var(--gray-primary)",
-  borderRadius: "20px",
-  "&:hover": {
-    transition: "all ease-in-out 0.2s",
-    backgroundColor: "var(--gray-secondary)",
-  },
 });
